IHRWEB-18713-ResultGeneration

+++ app/src/main/java/asvsapp/igloo/iglooGoldenData.java
@@ -2,8 +2,8 @@ package asvsapp.igloo;
 
 public class iglooGoldenData {
     
-    public float rank;
-    public float total_occurrences;
+    public int rank;
+    public int total_occurrences;
     public String search_term;
     public ssr ssr;
     public qsr qsr;
